backend/requirements.txt
asgiref==3.11.1
autobahn==25.12.2
cbor2==5.8.0
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
Pillow==12.1.1
psycopg2-binary==2.9.11
redis==7.2.0
sqlparse==0.5.5
twisted==25.5.0
whitenoise==6.11.0
